// src/submodule/common/interfaces/interfaces_styles.tsx
export interface MarginProps {
  no: any;
  horizontal0: any;
  vertivalNo: any;
  horizontal5: any;
  horizontal10: any;
  horizontal15: any;
  horizontal20: any;
  vertical5: any;
  vertical10: any;
  vertical15: any;
  vertical20: any;
  vertical25: any;
  vertical30: any;
  vertical35: any;
  vertical40: any;
  top5: any;
  top10: any;
  top15: any;
  top20: any;
  top25: any;
  top30: any;
  top35: any;
  top40: any;
  top45: any;
  top50: any;
  top55: any;
  top60: any;
  bottom5: any;
  bottom10: any;
  bottom15: any;
  bottom20: any;
  bottom25: any;
  bottom30: any;
  bottom35: any;
  bottom40: any;
  bottom45: any;
  bottom50: any;
  bottom55: any;
  bottom60: any;
}

export interface PaddingProps {
  no: any;
  horizontalNo: any;
  vertivalNo: any;
  horizontalFive: any;
  horizontalTen: any;
  horizontalFifteen: any;
  horizontalTwenty: any;
  horizontalThirty: any;
  horizontalForty: any;
  verticalFive: any;
  verticalTen: any;
  verticalFifteen: any;
  verticalTwenty: any;
  five: any;
  eight: any;
  ten: any;
  fifteen: any;
  twenty: any;
  twentyfive: any;
  top0: any;
  top5: any;
  top10: any;
  top15: any;
  top20: any;
  top25: any;
  top30: any;
  top35: any;
  bottom0: any;
  bottom5: any;
  bottom10: any;
  bottom15: any;
  bottom20: any;
  bottom25: any;
  bottom30: any;
  bottom35: any;
  right0: any;
  right5: any;
  right10: any;
  right15: any;
  right20: any;
  right25: any;
  right30: any;
  right35: any;
  left0: any;
  left5: any;
  left10: any;
  left15: any;
  left20: any;
  left25: any;
  left30: any;
  left35: any;
}

export interface ShadowInterface {
  input: any;
  box?: any;
}

export interface FontWeightInterface {
  heading: any;
  body: any;
  bold: any;
  light: any;
}

export interface ButtonInterface {
  wrap: any;
  common: any;
  primary: any;
  primaryColorFont: any;
  fluidWrap: any;
  fluid: any;
  basic: any;
  mini: any;
  tiny: any;
  small: any;
  default: any;
  large: any;
  big: any;
  miniFont: any;
  tinyFont: any;
  smallFont: any;
  defaultFont: any;
  largeFont: any;
  bigFont: any;
  commonText: any;
  commonMiniText: any;
  basicFont: any;
  whiteColor: any;
  whiteColorFont: any;
  bc1LightColor: any;
  bc1LightColorFont: any;
  bc1LighterColor: any;
  bc1LighterColorFont: any;
  bc1LightestColor: any;
  bc1LightestColorFont: any;
  spinnerWrap: any;
  disabled: any;
}

export interface HeadingProps {
  one: any;
  two: any;
  three: any;
  four: any;
  five: any;
  six: any;
}